feat(dal,sdf,app): Merge the concepts of Schema Variant and Definition
As we are moving away from the concept of SchemaVariantDefinitions, we want to move the metadata that was part of a SchemaVariantDefinition to be part of the SchemaVariant

This will allow us to start bringing back the authoring workflow without the need to have the concept of a SchemaVariantDefinition
