+++ lib/dal/src/installed_pkg/asset.rs
@@ -2,8 +2,6 @@ use super::{InstalledPkgId, InstalledPkgResult};
 use serde::{Deserialize, Serialize};
 use telemetry::prelude::*;
 
-// use crate::schema::variant::definition::SchemaVariantDefinitionId;
-use crate::schema::variant::definition::SchemaVariantDefinitionId;
 use crate::{
     impl_standard_model, pk, standard_model, standard_model_accessor, DalContext, FuncId, SchemaId,
     SchemaVariantId, StandardModel, Tenancy, Timestamp, Visibility,
@@ -40,7 +38,6 @@ pub enum InstalledPkgAssetKind {
     Func,
     Schema,
     SchemaVariant,
-    SchemaVariantDefinition,
 }
 
 /// An `InstalledPkgAsset` is a record of the installation of a package asset. It tracks the
@@ -83,12 +80,6 @@ pub enum InstalledPkgAssetTyped {
         installed_pkg_asset_id: InstalledPkgAssetId,
         installed_pkg_id: InstalledPkgId,
         id: SchemaVariantId,
-        hash: String,
-    },
-    SchemaVariantDefinition {
-        installed_pkg_asset_id: InstalledPkgAssetId,
-        installed_pkg_id: InstalledPkgId,
-        id: SchemaVariantDefinitionId,
         hash: String,
     },
 }
@@ -168,12 +159,6 @@ impl From<&InstalledPkgAsset> for InstalledPkgAssetTyped {
                 id: Into::<ulid::Ulid>::into(value.asset_id()).into(),
                 hash,
             },
-            InstalledPkgAssetKind::SchemaVariantDefinition => Self::SchemaVariantDefinition {
-                installed_pkg_asset_id,
-                installed_pkg_id,
-                id: Into::<ulid::Ulid>::into(value.asset_id()).into(),
-                hash,
-            },
         }
     }
 }
@@ -219,17 +204,6 @@ impl InstalledPkgAsset {
                 Into::<ulid::Ulid>::into(id).into(),
                 hash,
                 InstalledPkgAssetKind::SchemaVariant,
-            ),
-            InstalledPkgAssetTyped::SchemaVariantDefinition {
-                installed_pkg_id,
-                id,
-                hash,
-                ..
-            } => (
-                installed_pkg_id,
-                Into::<ulid::Ulid>::into(id).into(),
-                hash,
-                InstalledPkgAssetKind::SchemaVariantDefinition,
             ),
             InstalledPkgAssetTyped::Func {
                 installed_pkg_id,
@@ -277,14 +251,6 @@ impl InstalledPkgAsset {
                 installed_pkg_asset_id,
                 InstalledPkgAssetKind::Schema,
                 InstalledPkgAssetKind::SchemaVariant,
-            )),
-            InstalledPkgAssetTyped::SchemaVariantDefinition {
-                installed_pkg_asset_id,
-                ..
-            } => Err(super::InstalledPkgError::InstalledPkgKindMismatch(
-                installed_pkg_asset_id,
-                InstalledPkgAssetKind::Schema,
-                InstalledPkgAssetKind::SchemaVariantDefinition,
             )),
             InstalledPkgAssetTyped::Func {
                 installed_pkg_asset_id,
@@ -344,14 +310,7 @@ impl InstalledPkgAsset {
                 InstalledPkgAssetKind::SchemaVariant,
                 InstalledPkgAssetKind::Schema,
             )),
-            InstalledPkgAssetTyped::SchemaVariantDefinition {
-                installed_pkg_asset_id,
-                ..
-            } => Err(super::InstalledPkgError::InstalledPkgKindMismatch(
-                installed_pkg_asset_id,
-                InstalledPkgAssetKind::SchemaVariant,
-                InstalledPkgAssetKind::SchemaVariantDefinition,
-            )),
+
             InstalledPkgAssetTyped::Func {
                 installed_pkg_asset_id,
                 ..
@@ -375,14 +334,6 @@ impl InstalledPkgAsset {
                 installed_pkg_asset_id,
                 InstalledPkgAssetKind::Func,
                 InstalledPkgAssetKind::Schema,
-            )),
-            InstalledPkgAssetTyped::SchemaVariantDefinition {
-                installed_pkg_asset_id,
-                ..
-            } => Err(super::InstalledPkgError::InstalledPkgKindMismatch(
-                installed_pkg_asset_id,
-                InstalledPkgAssetKind::Func,
-                InstalledPkgAssetKind::SchemaVariantDefinition,
             )),
             InstalledPkgAssetTyped::SchemaVariant {
                 installed_pkg_asset_id,
