+++ lib/sdf-server/src/server/service/variant_definition/list_variant_defs.rs
@@ -1,9 +1,10 @@
-use super::SchemaVariantDefinitionResult;
 use crate::server::extract::{AccessBuilder, HandlerContext, PosthogClient};
 use crate::server::tracking::track;
+use crate::service::variant_definition::SchemaVariantResult;
 use axum::extract::OriginalUri;
 use axum::{extract::Query, Json};
-use dal::{schema::variant::definition::SchemaVariantDefinitionView, Visibility};
+use dal::schema::variant::SchemaVariantMetadataView;
+use dal::Visibility;
 use serde::{Deserialize, Serialize};
 
 #[derive(Deserialize, Serialize, Debug)]
@@ -16,7 +17,7 @@ pub struct ListVariantDefsRequest {
 #[derive(Deserialize, Serialize, Debug)]
 #[serde(rename_all = "camelCase")]
 pub struct ListVariantDefsResponse {
-    pub variant_defs: Vec<SchemaVariantDefinitionView>,
+    pub variant_defs: Vec<SchemaVariantMetadataView>,
 }
 
 pub async fn list_variant_defs(
@@ -25,10 +26,10 @@ pub async fn list_variant_defs(
     PosthogClient(posthog_client): PosthogClient,
     OriginalUri(original_uri): OriginalUri,
     Query(request): Query<ListVariantDefsRequest>,
-) -> SchemaVariantDefinitionResult<Json<ListVariantDefsResponse>> {
+) -> SchemaVariantResult<Json<ListVariantDefsResponse>> {
     let ctx = builder.build(request_ctx.build(request.visibility)).await?;
 
-    let schema_variant_definition_views = SchemaVariantDefinitionView::list(&ctx).await?;
+    let schema_variant_definition_views = SchemaVariantMetadataView::list(&ctx).await?;
 
     track(
         &posthog_client,
